Basic Agent: Almost Finished Coding Right-side algorithm
Identified the Six patches in the left side of the Grayscale image that are most similar to a given Grayscale value in the right side of the Grayscale image. Now, for each of the six identified patches, we must take the color representative of the re-colored middle pixel from the Left-side algorithm. If there is a majority representative color, use that color to re-color the middle pixel of your current patch. If there is no majority representative color, or there is a tie, color the middle pixel of the current patchthe same as the middle pixel color of the most similar test data patch.